init + apply test for aws-redis-node
This is our first test that uses one of our CI VPCs. The ids needed for this (vpc, subnets, etc) are being passed in as environment variables.

My thought is that if others want to fork and run the tests in their own infra, they can just set those env variables.