aws-iam-instance-profile Add option to disable attaching SSM policy (#77)
aws-iam-instance-profile Add option to disable attaching SSM policyThis PR adds the ability to disable attaching the SSM policy to the role created in aws-iam-instance-profile. It is believed this will disable the ssm-agent, and specifically prevent it from creating the ssm-user on the local machine, which has been messing up some pre-set UIDs.