adding in route53 Read Only Access (#96)
Adding in Route53ReadOnly Access to Route53Domains-PowerUser RoleNeeded to add the route53ReadOnly policy to the Route53 power user IAM role that I created previously. I need to be able to read and download the zone files for our hosted domains and this should give me the privileges to do that.