allow aws-s3-private-bucket take in custom lifecycle_rule (#173)
### Summary
This change allows an implementation of `aws-s3-private-bucket` to use the current lifecycle_rule, or provide a different one via a TF variable. For example, in my implementation I'd like to have a shorter 30-day TTL on "current" objects in the bucket. 

This is a little cumbersome, but I'm unsure if there's a cleaner way (new to TF lol) of doing this, since the `s3` resource doesn't take in a named argument but rather a configuration block. 

### Test Plan
Verified via `make plan` on a resource. 

### References
(Optional) Additional links to provide more context.