[fix] move build to github actions (#201)
travis-ci has proven to be too unreliable and difficult to maintain. It seems like GHA will be a better thing for us to build on.

I toyed with a few approaches, but decided to run each module as a separate job. The setup time is pretty fast and having a separate status check for each module is pretty nice.

### Test Plan
* CI

### References
*