Allow lambda@edge in aws-lambda-function module (#242)
I want to use this module to create lambdas for cloudfront. That
requires supporting lambda@Edge. The two main changes here are:
1. Adding edgelambda as an authorized service for the role
2. granting permission create/update permission on * log groups.

The second is I suppose slightly controversial. because its an edge
lambda, we would have to precreate the log groups in every single aws
region, and I have no clue how to do that. I think this is an okay
tradeoff for now, and tried to document that in the description.

Test Plan:
wait for CI
use it in stacked TF and see if it works.