[feature] domain redirects via Cloudfront (#243)
There should be a canonical way to redirect from a naked domain like
http://foo.com or https://foo.com to https://www.foo.com, while
including the appropriate security headers like HSTS.

This is a first pass at such an abstraction.

I separated out the security-headers lambda since it is independently
useful in other cloudfront distribution settings. It's not very
configurable yet, and doesn't quite live up to the promise of its name,
but perhaps will get there in time.

Test Plan:
push the PR so I can try using it in another repo.