Add HSTS headers to aws-single-page-static-site (#244)
This module already enforces that the site is accessed with HTTPS via
the cloudfront `"redirect-to-https"` policy.

Since that's true, we can also set HSTS headers on all the responses
using a cloudfront lambda, that, I've conveniently created already in
https://github.com/chanzuckerberg/cztack/pull/243

This does that.

Test Plan:
push up PR
update connections-infra module ref to new commit
make apply
curl to see if it works