Fix permanently dirty plan with aws-lambda-edge-security-headers (#253)
This was using the wrong hash function for the lambda source_code_hash.
THis means that every plan suggested a change, but no change actually
happened. The only problem was that the hash function was different, so
the comparison always had a difference.

Test Plan: plan is clean when using this ref